Move Farming Task classes to new package
Changes:
- Create new task subpackage under Farming
- Move all related Task classes under the new task subpackage
- Move tests as well

ref: #1298
