+++ mars-sim-core/src/main/java/com/mars_sim/core/building/function/farming/task/TendAlgaePondMeta.java
@@ -4,7 +4,7 @@
  * @date 2023-09-19
  * @author Manny Kung
  */
-package com.mars_sim.core.building.function.task;
+package com.mars_sim.core.building.function.farming.task;
 
 import java.util.ArrayList;
 import java.util.List;
