+++ mars-sim-core/src/main/java/com/mars_sim/core/building/function/farming/task/TendHousekeeping.java
@@ -4,7 +4,7 @@
  * @date 2024-06-02
  * @author Barry Evans
  */
-package com.mars_sim.core.building.function.task;
+package com.mars_sim.core.building.function.farming.task;
 
 
 import com.mars_sim.core.building.function.HouseKeeping;
