+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/util/MetaTaskUtil.java
@@ -17,12 +17,12 @@ import java.util.stream.Collectors;
 import com.mars_sim.core.Simulation;
 import com.mars_sim.core.activities.GroupActivityMetaTask;
 import com.mars_sim.core.building.function.cooking.task.CookMealMeta;
+import com.mars_sim.core.building.function.farming.task.TendAlgaePondMeta;
+import com.mars_sim.core.building.function.farming.task.TendFishTankMeta;
+import com.mars_sim.core.building.function.farming.task.TendGreenhouseMeta;
 import com.mars_sim.core.building.function.task.ObserveAstronomicalObjectsMeta;
 import com.mars_sim.core.building.function.task.OptimizeSystemMeta;
 import com.mars_sim.core.building.function.task.ProduceFoodMeta;
-import com.mars_sim.core.building.function.task.TendAlgaePondMeta;
-import com.mars_sim.core.building.function.task.TendFishTankMeta;
-import com.mars_sim.core.building.function.task.TendGreenhouseMeta;
 import com.mars_sim.core.building.function.task.ToggleFuelPowerSourceMeta;
 import com.mars_sim.core.building.task.MaintainBuildingMeta;
 import com.mars_sim.core.manufacture.task.ManufacturingMetaTask;
